Replace miners multiple_select with a text_input
multiple_select assumes values are already present for a run and you
just choose which ones you'd like to select.

Having a text_input that splits words into the array we need in the db
is hacky for multiple reasons:
1. Users need to know how to input values in this field
2. Inputs need to be altered both when entered and when retrieved

But at the same time it is an easy way to alter this data without
touching any frontend code.